+++ src/model/gameboard.h
@@ -5,10 +5,10 @@
 #include <vector>
 
 namespace GameBoardConstants {
-    static const uint8_t FREE_TILE_ID = 1; 
-    static const uint8_t NO_SELECTED_INDEX = 255;
-    static const uint8_t MIN_TILE_TYPES = 2;
-    static const uint8_t MAX_TILE_TYPES = 100;
+    static const unsigned int FREE_TILE_ID = 1; 
+    static const unsigned int NO_SELECTED_INDEX = 255;
+    static const unsigned int MIN_TILE_TYPES = 2;
+    static const unsigned int MAX_TILE_TYPES = 100;
     static const unsigned int MAX_WIDTH = 15;
     static const unsigned int MAX_HEIGHT = 15;
     static const unsigned int MIN_WIDTH = 2;
@@ -26,27 +26,27 @@ class GameBoard {
 
     public:
 
-        GameBoard(const unsigned int& newWidth, const unsigned int& newHeight, const uint8_t& newTileTypes);
+        GameBoard(const unsigned int& newWidth, const unsigned int& newHeight, const unsigned int& newTileTypes);
         GameBoard();
         // Accessors
 
-        inline const uint8_t& getMatchesNeeded() const { return matchesNeeded; }
+        inline const unsigned int& getMatchesNeeded() const { return matchesNeeded; }
         inline const GameBoardDimensions& getBoardDimensions() const { return boardDimensions; }
         inline const std::vector<TileData>& getTiles() const { return tiles; }
-        inline const uint8_t& getNumTileTypes() const { return numTileTypes; }
+        inline const unsigned int& getNumTileTypes() const { return numTileTypes; }
 
         // Mutators
 
-        void setTileFlipped(const uint8_t& index) {
+        void setTileFlipped(const unsigned int& index) {
             tiles[index].flags = TileFlags::FLIPPED;
         }
 
-        void setTilesMatched(const uint8_t& index1, const uint8_t& index2) {
+        void setTilesMatched(const unsigned int& index1, const unsigned int& index2) {
             tiles[index1].flags = TileFlags::MATCHED;
             tiles[index2].flags = TileFlags::MATCHED;
         }      
 
-        void setTilesUnflipped(const uint8_t& index1, const uint8_t& index2) {
+        void setTilesUnflipped(const unsigned int& index1, const unsigned int& index2) {
             tiles[index1].flags = TileFlags::UNFLIPPED;
             tiles[index2].flags = TileFlags::UNFLIPPED;
         }
@@ -59,8 +59,8 @@ class GameBoard {
 
         GameBoardDimensions boardDimensions;
 
-        uint8_t numTileTypes;
-        uint8_t matchesNeeded;
+        unsigned int numTileTypes;
+        unsigned int matchesNeeded;
 
         std::vector<TileData> tiles;
